Move mongo tests to testapps, integrate testapps-build plugin 0.0.10, remove obselete plugins from the main pom
