--- foundation/org.eclipse.persistence.nosql/src/it/resources/META-INF/persistence.xml
@@ -1,47 +1,0 @@
-<!--
-
-    Copyright (c) 2018, 2019 Oracle and/or its affiliates. All rights reserved.
-
-    This program and the accompanying materials are made available under the
-    terms of the Eclipse Public License v. 2.0 which is available at
-    http://www.eclipse.org/legal/epl-2.0,
-    or the Eclipse Distribution License v. 1.0 which is available at
-    http://www.eclipse.org/org/documents/edl-v10.php.
-
-    SPDX-License-Identifier: EPL-2.0 OR BSD-3-Clause
-
--->
-
-<persistence xmlns="http://java.sun.com/xml/ns/persistence" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://java.sun.com/xml/ns/persistence persistence_1_0.xsd" version="1.0">
-    <persistence-unit name="mongo" transaction-type="RESOURCE_LOCAL">
-        <provider>org.eclipse.persistence.jpa.PersistenceProvider</provider>
-        <class>org.eclipse.persistence.testing.models.jpa.mongo.Order</class>
-        <class>org.eclipse.persistence.testing.models.jpa.mongo.Customer</class>
-        <class>org.eclipse.persistence.testing.models.jpa.mongo.Buyer</class>
-        <class>org.eclipse.persistence.testing.models.jpa.mongo.LineItem</class>
-        <class>org.eclipse.persistence.testing.models.jpa.mongo.Address</class>
-        <properties>
-            <property name="eclipselink.target-database" value="org.eclipse.persistence.nosql.adapters.mongo.MongoPlatform"/>
-            <property name="eclipselink.nosql.connection-spec" value="org.eclipse.persistence.nosql.adapters.mongo.MongoConnectionSpec"/>
-        </properties>
-    </persistence-unit>
-    <persistence-unit name="mongo-xml" transaction-type="RESOURCE_LOCAL">
-        <provider>org.eclipse.persistence.jpa.PersistenceProvider</provider>
-        <mapping-file>META-INF/mongo-orm.xml</mapping-file>
-        <properties>
-            <property name="eclipselink.orm.validate.schema" value="true"/>
-            <property name="eclipselink.target-database" value="org.eclipse.persistence.nosql.adapters.mongo.MongoPlatform"/>
-            <property name="eclipselink.nosql.connection-spec" value="org.eclipse.persistence.nosql.adapters.mongo.MongoConnectionSpec"/>
-        </properties>
-    </persistence-unit>
-    <persistence-unit name="mongo-dynamic" transaction-type="RESOURCE_LOCAL">
-        <provider>org.eclipse.persistence.jpa.PersistenceProvider</provider>
-        <mapping-file>META-INF/mongo-dynamic.xml</mapping-file>
-        <properties>
-            <property name="eclipselink.orm.validate.schema" value="true"/>
-            <property name="eclipselink.weaving" value="static"/>
-            <property name="eclipselink.target-database" value="org.eclipse.persistence.nosql.adapters.mongo.MongoPlatform"/>
-            <property name="eclipselink.nosql.connection-spec" value="org.eclipse.persistence.nosql.adapters.mongo.MongoConnectionSpec"/>
-        </properties>
-    </persistence-unit>
-</persistence>
