--- jpa/eclipselink.jpa.nosql.test/src/it/assembly/assembly-client.xml
@@ -1,85 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-
-    Copyright (c) 2019 Oracle and/or its affiliates. All rights reserved.
-
-    This program and the accompanying materials are made available under the
-    terms of the Eclipse Public License v. 2.0 which is available at
-    http://www.eclipse.org/legal/epl-2.0,
-    or the Eclipse Distribution License v. 1.0 which is available at
-    http://www.eclipse.org/org/documents/edl-v10.php.
-
-    SPDX-License-Identifier: EPL-2.0 OR BSD-3-Clause
-
--->
-
-<assembly xmlns="http://maven.apache.org/ASSEMBLY/2.0.0"
-          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-          xsi:schemaLocation="http://maven.apache.org/ASSEMBLY/2.0.0 http://maven.apache.org/xsd/assembly-2.0.0.xsd">
-    <id>${item}_client</id>
-    <formats>
-        <format>jar</format>
-    </formats>
-    <includeBaseDirectory>false</includeBaseDirectory>
-    <componentDescriptors>
-        <componentDescriptor>${integration.test.directory}/assembly/${item}.xml</componentDescriptor>
-    </componentDescriptors>
-    <dependencySets>
-        <dependencySet>
-            <outputDirectory>.</outputDirectory>
-            <includes>
-                <include>org.eclipse.persistence:org.eclipse.persistence.core.test.framework</include>
-            </includes>
-            <scope>test</scope>
-            <unpack>true</unpack>
-            <unpackOptions>
-                <includes>
-                    <include>org/eclipse/persistence/testing/framework/junit/*.class</include>
-                </includes>
-            </unpackOptions>
-        </dependencySet>
-        <dependencySet>
-            <outputDirectory>.</outputDirectory>
-            <includes>
-                <include>org.eclipse.persistence:org.eclipse.persistence.jpa.test.framework</include>
-            </includes>
-            <scope>test</scope>
-            <unpack>true</unpack>
-            <unpackOptions>
-                <includes>
-                    <include>org/eclipse/persistence/testing/framework/junit/*.class</include>
-                    <include>org/eclipse/persistence/testing/framework/jpa/server/*.class</include>
-                </includes>
-            </unpackOptions>
-        </dependencySet>
-        <dependencySet>
-            <outputDirectory>.</outputDirectory>
-            <includes>
-                <include>org.eclipse.persistence:org.eclipse.persistence.jpa.test:test-jar</include>
-            </includes>
-            <scope>test</scope>
-            <unpack>true</unpack>
-            <unpackOptions>
-                <includes>
-                    <include>org/eclipse/persistence/testing/tests/jpa/${packageName}/**/*.class</include>
-                </includes>
-            </unpackOptions>
-        </dependencySet>
-    </dependencySets>
-    <files>
-        <file>
-            <source>${project.build.testOutputDirectory}/server/jndi.properties</source>
-            <outputDirectory>.</outputDirectory>
-            <filtered>true</filtered>
-        </file>
-    </files>
-    <fileSets>
-        <fileSet>
-            <directory>${project.build.testOutputDirectory}/server</directory>
-            <outputDirectory>META-INF</outputDirectory>
-            <includes>
-                <include>application-client.xml</include>
-            </includes>
-        </fileSet>
-    </fileSets>
-</assembly>
