--- jpa/eclipselink.jpa.nosql.test/src/it/resources/eclipselink-mongo-model-xml/persistence.xml
@@ -1,29 +1,0 @@
-<!--
-
-    Copyright (c) 2018, 2020 Oracle and/or its affiliates. All rights reserved.
-
-    This program and the accompanying materials are made available under the
-    terms of the Eclipse Public License v. 2.0 which is available at
-    http://www.eclipse.org/legal/epl-2.0,
-    or the Eclipse Distribution License v. 1.0 which is available at
-    http://www.eclipse.org/org/documents/edl-v10.php.
-
-    SPDX-License-Identifier: EPL-2.0 OR BSD-3-Clause
-
--->
-
-<persistence xmlns="http://java.sun.com/xml/ns/persistence" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://java.sun.com/xml/ns/persistence persistence_1_0.xsd" version="1.0">
-    <persistence-unit name="@default@" transaction-type="@transaction-type@">
-        <provider>org.eclipse.persistence.jpa.PersistenceProvider</provider>
-        <mapping-file>META-INF/mongo-orm.xml</mapping-file>
-        <properties>
-            <property name="eclipselink.orm.validate.schema" value="true"/>
-            <property name="eclipselink.target-server" value="@server-platform@"/>
-            <property name="eclipselink.target-database" value="org.eclipse.persistence.nosql.adapters.mongo.MongoPlatform"/>
-            <property name="eclipselink.nosql.connection-spec" value="org.eclipse.persistence.nosql.adapters.mongo.MongoConnectionSpec"/>
-            <!--property name="eclipselink.nosql.property.mongo.port" value="localhost"/>
-            <property name="eclipselink.nosql.property.mongo.host" value="localhost"/>
-            <property name="eclipselink.nosql.property.mongo.db" value="mydb"/-->
-        </properties>
-    </persistence-unit>
-</persistence>
