--- jpa/eclipselink.jpa.nosql.test/src/it/resources/server/application-client.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0"?>
-<!DOCTYPE application-client PUBLIC "-//Sun Microsystems, Inc.//DTD J2EE Application Client 1.2//EN" "http://java.sun.com/j2ee/dtds/application-client_1_2.dtd">
-<!--
-
-    Copyright (c) 2018, 2022 Oracle and/or its affiliates. All rights reserved.
-
-    This program and the accompanying materials are made available under the
-    terms of the Eclipse Public License v. 2.0 which is available at
-    http://www.eclipse.org/legal/epl-2.0,
-    or the Eclipse Distribution License v. 1.0 which is available at
-    http://www.eclipse.org/org/documents/edl-v10.php.
-
-    SPDX-License-Identifier: EPL-2.0 OR BSD-3-Clause
-
--->
-
-<application-client>
-    <display-name>servertest</display-name>
-    <ejb-ref>
-        <ejb-ref-name>ejb/TestRunner</ejb-ref-name>
-        <ejb-ref-type>Session</ejb-ref-type>
-        <remote>org.eclipse.persistence.testing.framework.jpa.server.TestRunner</remote>
-    </ejb-ref>
-</application-client>
