--- jpa/eclipselink.jpa.nosql.test/src/it/resources/server/ejb-jar.xml
@@ -1,34 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-
-    Copyright (c) 2018, 2020 Oracle and/or its affiliates. All rights reserved.
-
-    This program and the accompanying materials are made available under the
-    terms of the Eclipse Public License v. 2.0 which is available at
-    http://www.eclipse.org/legal/epl-2.0,
-    or the Eclipse Distribution License v. 1.0 which is available at
-    http://www.eclipse.org/org/documents/edl-v10.php.
-
-    SPDX-License-Identifier: EPL-2.0 OR BSD-3-Clause
-
--->
-
-<ejb-jar xmlns="http://java.sun.com/xml/ns/javaee"
-          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-          xsi:schemaLocation="http://java.sun.com/xml/ns/javaee
-          http://java.sun.com/xml/ns/javaee/ejb-jar_3_0.xsd"
-          version="3.0">
-    <enterprise-beans>
-        <session>
-            <ejb-name>TestRunner</ejb-name>
-            <persistence-context-ref>
-                <persistence-context-ref-name>persistence/@default@/entity-manager</persistence-context-ref-name>
-                <persistence-unit-name>@default@</persistence-unit-name>
-            </persistence-context-ref>
-            <persistence-unit-ref>
-                <persistence-unit-ref-name>persistence/@default@/factory</persistence-unit-ref-name>
-                <persistence-unit-name>@default@</persistence-unit-name>
-            </persistence-unit-ref>
-        </session>
-    </enterprise-beans>
-</ejb-jar>
