--- jpa/eclipselink.jpa.nosql.test/src/it/resources/server/persistence_cachecoordination.xml
@@ -1,37 +1,0 @@
-<?xml version="1.0" encoding="windows-1252" ?>
-<!--
-
-    Copyright (c) 2018, 2020 Oracle and/or its affiliates. All rights reserved.
-
-    This program and the accompanying materials are made available under the
-    terms of the Eclipse Public License v. 2.0 which is available at
-    http://www.eclipse.org/legal/epl-2.0,
-    or the Eclipse Distribution License v. 1.0 which is available at
-    http://www.eclipse.org/org/documents/edl-v10.php.
-
-    SPDX-License-Identifier: EPL-2.0 OR BSD-3-Clause
-
--->
-
-<persistence xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-             xsi:schemaLocation="http://java.sun.com/xml/ns/persistence http://java.sun.com/xml/ns/persistence/persistence_1_0.xsd"
-             version="1.0" xmlns="http://java.sun.com/xml/ns/persistence">
-    <persistence-unit name="@default@">
-        <provider>org.eclipse.persistence.jpa.PersistenceProvider</provider>
-        <@datasource-type@>@data-source-name@</@datasource-type@>
-        <class>org.eclipse.persistence.testing.models.jpa.fieldaccess.advanced.Employee</class>
-        <class>org.eclipse.persistence.testing.models.jpa.fieldaccess.advanced.Address</class>
-        <properties>
-            <property name="eclipselink.target-server" value="@server-platform@"/>
-            <property name="eclipselink.target-database" value="@database-platform@"/>
-            <property name="eclipselink.logging.logger" value="DefaultLogger"/>
-            <property name="eclipselink.logging.level" value="FINEST"/>
-
-            <property name="eclipselink.cache.type.default" value="Full" />
-            <property name="eclipselink.cache.coordination.protocol" value="@cachecoordination-protocol@" />
-            <property name="eclipselink.cache.coordination.jms.host" value="@topic-host-url@" />
-            <property name="eclipselink.cache.coordination.jms.topic" value="@topic-name@" />
-            <property name="eclipselink.cache.coordination.jms.factory" value="@topic-connection-factory@" />
-        </properties>
-    </persistence-unit>
-</persistence>
