--- jpa/eclipselink.jpa.nosql.test/src/it/resources/wildfly/elResources.xml
@@ -1,23 +1,0 @@
-<!--
-
-    Copyright (c) 2020 Oracle and/or its affiliates. All rights reserved.
-
-    This program and the accompanying materials are made available under the
-    terms of the Eclipse Public License v. 2.0 which is available at
-    http://www.eclipse.org/legal/epl-2.0,
-    or the Eclipse Distribution License v. 1.0 which is available at
-    http://www.eclipse.org/org/documents/edl-v10.php.
-
-    SPDX-License-Identifier: EPL-2.0 OR BSD-3-Clause
-
--->
-
-<resources>
-<resource-root path="${org.eclipse.persistence:org.eclipse.persistence.antlr:jar}"/>
-<resource-root path="${org.eclipse.persistence:org.eclipse.persistence.asm:jar}"/>
-<resource-root path="${org.eclipse.persistence:org.eclipse.persistence.jpa.jpql:jar}"/>
-<resource-root path="${org.eclipse.persistence:org.eclipse.persistence.core:jar}"/>
-<resource-root path="${org.eclipse.persistence:org.eclipse.persistence.nosql:jar}"/>
-<resource-root path="${org.eclipse.persistence:org.eclipse.persistence.jpa:jar}"/>
-<resource-root path="${org.eclipse.persistence:org.eclipse.persistence.moxy:jar}"/>
-</resources>
